#95 - refactor: Service 단위 테스트 클래스 패키지 변경
