+++ src/test/java/com/gxdxx/instagram/domain/user/application/UserCreateServiceTest.java
@@ -1,4 +1,4 @@
-package com.gxdxx.instagram.service.user;
+package com.gxdxx.instagram.domain.user.application;
 
 import com.gxdxx.instagram.domain.user.application.UserCreateService;
 import com.gxdxx.instagram.domain.user.dao.RoleRepository;
