+++ src/test/java/com/gxdxx/instagram/domain/user/application/UserProfileUpdateServiceTest.java~
@@ -1,0 +1,144 @@
+package com.gxdxx.instagram.domain.user;
+
+import com.gxdxx.instagram.domain.user.application.UserProfileUpdateService;
+import com.gxdxx.instagram.domain.user.dto.request.UserProfileUpdateRequest;
+import com.gxdxx.instagram.domain.user.dto.response.UserProfileUpdateResponse;
+import com.gxdxx.instagram.domain.user.domain.User;
+import com.gxdxx.instagram.domain.user.exception.NicknameAlreadyExistsException;
+import com.gxdxx.instagram.domain.user.exception.UserNotFoundException;
+import com.gxdxx.instagram.domain.user.dao.UserRepository;
+import com.gxdxx.instagram.global.config.s3.S3Uploader;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.mock.web.MockMultipartFile;
+
+import java.util.Optional;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.Mockito.*;
+
+@ExtendWith(MockitoExtension.class)
+public class UserProfileUpdateServiceTest {
+
+    @InjectMocks
+    private UserProfileUpdateService userProfileUpdateService;
+    @Mock
+    private UserRepository userRepository;
+    @Mock
+    private S3Uploader s3Uploader;
+
+    private static final String NICKNAME = "nickname";
+    private static final String NEW_NICKNAME = "newNickname";
+    private static final String DUPLICATED_NICKNAME = "duplicatedNickname";
+    private static final String ENCODED_PASSWORD = "encodedPassword";
+    private static final String PROFILE_IMAGE_URL = "profileImageUrl";
+    private static final String NEW_PROFILE_IMAGE_URL = "newProfileImageUrl";
+    private static final String DIR_NAME = "images";
+
+    @Test
+    @DisplayName("[프로필 수정] - 성공 (기존 닉네임과 같은 경우)")
+    void updateUserProfile_WithSameNickname_ShouldSucceed() {
+        // given
+        UserProfileUpdateRequest request = createUserProfileUpdateRequest(NICKNAME, getMockMultipartFile());
+        User savedUser = createSavedUser();
+
+        when(userRepository.findByNickname(request.nickname())).thenReturn(Optional.of(savedUser));
+        when(s3Uploader.upload(request.profileImage(), DIR_NAME)).thenReturn(NEW_PROFILE_IMAGE_URL);
+
+        // when
+        UserProfileUpdateResponse response = userProfileUpdateService.updateUserProfile(request, NICKNAME);
+
+        // then
+        assertEquals(request.nickname(), response.nickname());
+        assertEquals(NEW_PROFILE_IMAGE_URL, response.profileImageUrl());
+    }
+
+    @Test
+    @DisplayName("[프로필 수정] - 성공 (기존 닉네임과 다른 경우)")
+    void updateUserProfile_WithDifferentNickname_ShouldSucceed() {
+        // given
+        UserProfileUpdateRequest request = createUserProfileUpdateRequest(NEW_NICKNAME, getMockMultipartFile());
+        User savedUser = createSavedUser();
+
+        when(userRepository.findByNickname(request.nickname())).thenReturn(Optional.empty());
+        when(userRepository.findByNickname(savedUser.getNickname())).thenReturn(Optional.of(savedUser));
+        when(s3Uploader.upload(request.profileImage(), DIR_NAME)).thenReturn(NEW_PROFILE_IMAGE_URL);
+
+        // when
+        UserProfileUpdateResponse response = userProfileUpdateService.updateUserProfile(request, savedUser.getNickname());
+
+        // then
+        assertEquals(request.nickname(), response.nickname());
+        assertEquals(NEW_PROFILE_IMAGE_URL, response.profileImageUrl());
+    }
+
+    @Test
+    @DisplayName("[프로필 수정] - 실패 (이미 존재하는 닉네임)")
+    void updateUserProfile_withExistingNickname_shouldThrowException() {
+        // given
+        UserProfileUpdateRequest request = createUserProfileUpdateRequest(NEW_NICKNAME, getMockMultipartFile());
+        User duplicatedUser = createDuplicatedUser();
+
+        when(userRepository.findByNickname(request.nickname())).thenReturn(Optional.of(duplicatedUser));
+
+        // when & then
+        Assertions.assertThrows(NicknameAlreadyExistsException.class, () -> userProfileUpdateService.updateUserProfile(request, duplicatedUser.getNickname()));
+    }
+
+    @Test
+    @DisplayName("[프로필 수정] - 실패 (존재하지 않는 유저 && 기존 닉네임과 같은 경우)")
+    void updateUserProfile_withNonExistingUser_withSameNickname_shouldThrowUserNotFoundException() {
+        // given
+        UserProfileUpdateRequest request = createUserProfileUpdateRequest(NICKNAME, getMockMultipartFile());
+        User savedUser = createSavedUser();
+
+        when(userRepository.findByNickname(savedUser.getNickname())).thenReturn(Optional.empty());
+
+        // when & then
+        Assertions.assertThrows(UserNotFoundException.class, () -> userProfileUpdateService.updateUserProfile(request, savedUser.getNickname()));
+    }
+
+    @Test
+    @DisplayName("[프로필 수정] - 실패 (존재하지 않는 유저 && 기존 닉네임과 다른 경우)")
+    void updateUserProfile_withNonExistingUser_withDifferentNickname_shouldThrowUserNotFoundException() {
+        // given
+        UserProfileUpdateRequest request = createUserProfileUpdateRequest(NEW_NICKNAME, getMockMultipartFile());
+        User user = createSavedUser();
+
+        when(userRepository.findByNickname(request.nickname())).thenReturn(Optional.empty());
+        when(userRepository.findByNickname(user.getNickname())).thenReturn(Optional.empty());
+
+        // when & then
+        Assertions.assertThrows(UserNotFoundException.class, () -> userProfileUpdateService.updateUserProfile(request, user.getNickname()));
+    }
+
+    private User createSavedUser() {
+        return User.of(NICKNAME, ENCODED_PASSWORD, PROFILE_IMAGE_URL);
+    }
+
+    private User createDuplicatedUser() {
+        return User.of(DUPLICATED_NICKNAME, ENCODED_PASSWORD, PROFILE_IMAGE_URL);
+    }
+
+    private UserProfileUpdateRequest createUserProfileUpdateRequest(String updateNickname, MockMultipartFile profileImage) {
+        return new UserProfileUpdateRequest(updateNickname, profileImage);
+    }
+
+    private MockMultipartFile getMockMultipartFile() {
+        MockMultipartFile mockFile = new MockMultipartFile(
+                "file",
+                "test.png",
+                "image/png",
+                "test content".getBytes()
+        );
+        return mockFile;
+    }
+
+}
